Score the Michigan Retinal Degeneration Questionnaire

Score the Michigan Retinal Degeneration Questionnaire

Usage

scoreMRDQ(dat, verbose = 0)

Arguments

dat

data.frame with N rows and 60 columns. First column is patient ID and the remaining columns are responses to the 59 items. Responses should be integers from 0 to 5.

verbose

A number indicating the amount of printing during function execution. 0 (default) is none. Higher numbers may result in more printing.

Value

A list with two components, each an N by 8 matrix. The first, named thetas, contains the ID and 7 domain scores for each patient. The second, named ses, contains the ID and the standard errors of the thetas.

Examples

justme <- read.table(
  text = "Me,0,0,0,0,0,0,0,0,1,1,1,0,0,0,1,1,1,2,0,0,0,0,1,1,1,0,0,0,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,2,1,1,0,0,0,0,0,2,0,0,0",
  sep = ",")
scoreMRDQ(justme)
#> $thetas
#>   ID       CV      Col        Cnt         SF         PF        MF        PS
#> 1 Me -1.08406 -1.34802 0.06385294 -0.4440405 -0.6177632 -1.383663 -1.432964
#> 
#> $ses
#>   ID        CV      Col       Cnt        SF        PF       MF        PS
#> 1 Me 0.2925964 0.578973 0.2301165 0.1298078 0.3168943 0.339391 0.3869802
#> 

justyou <- data.frame(ID="You", matrix(nrow=1, sample(0:3, 59, TRUE)))
scoreMRDQ(justyou)
#> $thetas
#>    ID        CV         Col       Cnt         SF        PF         MF
#> 1 You 0.2685624 -0.02701186 0.5760405 -0.2741323 0.8834815 -0.2451467
#>           PS
#> 1 -0.4595375
#> 
#> $ses
#>    ID        CV       Col       Cnt       SF       PF        MF        PS
#> 1 You 0.1929196 0.4133385 0.2069117 0.174964 0.218055 0.1768086 0.2713215
#> 

names(justme) <- names(justyou) <- c("ID", sprintf("Q%0.2d", seq(59)))
justus <- rbind(justme, justyou)
scoreMRDQ(justus, verbose = 1)
#> i =  1 
#> i =  2 
#> $thetas
#>    ID         CV         Col        Cnt         SF         PF         MF
#> 1  Me -1.0840599 -1.34801952 0.06385294 -0.4440405 -0.6177632 -1.3836627
#> 2 You  0.2685624 -0.02701186 0.57604053 -0.2741323  0.8834815 -0.2451467
#>           PS
#> 1 -1.4329642
#> 2 -0.4595375
#> 
#> $ses
#>    ID        CV       Col       Cnt        SF        PF        MF        PS
#> 1  Me 0.2925964 0.5789730 0.2301165 0.1298078 0.3168943 0.3393910 0.3869802
#> 2 You 0.1929196 0.4133385 0.2069117 0.1749640 0.2180550 0.1768086 0.2713215
#>